pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.0",
    "flask>=3.1.0",
    "near-api-py>=0.1.0",
    "openai>=1.65.2",
    "python-dotenv>=1.0.1",
    "schedule>=1.2.2",
    "tweepy>=4.15.0",
    "twilio>=9.4.6",
]
